Poi3D for Android (GDX)
