+++ vtm-android-example/AndroidManifest.xml
@@ -47,6 +47,9 @@
             android:name=".GdxActivity"
             android:configChanges="keyboardHidden|orientation|screenSize" />
         <activity
+            android:name=".GdxPoi3DActivity"
+            android:configChanges="keyboardHidden|orientation|screenSize" />
+        <activity
             android:name=".JeoIndoorActivity"
             android:configChanges="keyboardHidden|orientation|screenSize" />
         <activity
