+++ vtm-gdx/assets/data/g3d/test.g3db
