+++ vtm-gdx/assets/data/g3d/tree.png
