+++ vtm-gdx/assets/data/g3d/treeA.g3dj
