+++ vtm-gdx/build.gradle
@@ -9,7 +9,7 @@ dependencies {
 
 sourceSets {
     main.java.srcDirs = ['src']
-    main.resources.srcDirs = ['src']
+    main.resources.srcDirs = ['src','assets']
 }
 
 if (project.hasProperty("SONATYPE_USERNAME")) {
