+++ vtm-gdx/src/org/oscim/gdx/poi3d/MapCamera.java
@@ -1,4 +1,4 @@
-package org.oscim.test.gdx.poi3d;
+package org.oscim.gdx.poi3d;
 
 import com.badlogic.gdx.graphics.Camera;
 import com.badlogic.gdx.math.Matrix4;
